requirements.txt
flask
flask-login
sqlalchemy
opencv-python
face_recognition
pandas
gunicorn
python-dotenv
